% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm6.R
\name{make_hs_classes}
\alias{make_hs_classes}
\title{Run model}
\usage{
make_hs_classes(df, nper)
}
\arguments{
\item{df}{school data frame used in run_model setup}

\item{nper}{number of school periods; defaults to 8}
}
\description{
Perform a single model run
}
