% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm6.R
\name{make_quarantine}
\alias{make_quarantine}
\title{Update quarantine}
\usage{
make_quarantine(
  class_quarantine,
  df.u,
  quarantine.length = 10,
  quarantine.grace = 3,
  hs = F,
  hs.classes = NA
)
}
\arguments{
\item{class_quarantine}{data frame of quarantine times}

\item{df.u}{data frame of infections whose classes should be quarantined}
}
\value{
class_quarantine updated
}
\description{
Mark classes for quarantine based on current symptomatic infections
}
