% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm6.R
\name{make_schedule}
\alias{make_schedule}
\title{Make schedule}
\usage{
make_schedule(time = 30, type = "base", total_days = 5, df)
}
\arguments{
\item{time}{number of days; defaults to 30}

\item{type}{"base", "On/off", "A/B", "Remote"; defaults to "base"}

\item{total_days}{number of days in school; defaults to 5}

\item{df}{data frame from make_school()}
}
\value{
d Returns a n x time data frame that indicates whether an individual is
in the school building at a particular time
}
\description{
Make a schedule of when individuals in the school community are
present/absent
}
