% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm6.R
\name{make_school}
\alias{make_school}
\title{Make school}
\usage{
make_school(
  synthpop = synthMaryland,
  n_other_adults = 30,
  includeFamily = F,
  n_class = 4
)
}
\arguments{
\item{synthpop}{synthetic population; defaults to synthMaryland stored in file}

\item{n_other_adults}{Number of adults in the school other than primary teachers; defaults to 30}

\item{includeFamily}{whether to include family and adult family members of teachers, default = FALSE}

\item{n_class}{number of classes per grade}
}
\value{
out data frame of child and teacher attributes
}
\description{
This function allows you to sort a synthetic population into classes.
It also assigns children to groups for alternating schedules and
ensures that children are in the same group as siblings.
It adds non-primary teacher staff, and if families are included, includes
two adult family members per child and one per adult staff member.
}
