% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm6.R
\name{mult_runs}
\alias{mult_runs}
\title{Run model multiple times and summarize results}
\usage{
mult_runs(
  N,
  n_other_adults,
  n_contacts,
  rel_trans_HH,
  rel_trans,
  rel_trans_CC,
  rel_trans_adult,
  p_asymp_adult,
  child_prob,
  adult_prob,
  p_asymp_child,
  attack,
  child_trans,
  child_susp,
  child_vax,
  p_subclin_adult,
  p_subclin_child,
  teacher_susp,
  disperse_transmission,
  n_staff_contact,
  n_HH,
  num_adults,
  family_susp,
  n_start,
  time_seed_inf,
  days_inf,
  mult_asymp,
  mult_asymp_child,
  seed_asymp,
  isolate,
  dedens,
  run_specials_now,
  time,
  notify,
  test,
  test_sens,
  test_frac,
  test_days,
  test_type,
  quarantine.length,
  quarantine.grace,
  type,
  total_days,
  includeFamily,
  synthpop,
  class,
  n_class,
  high_school,
  nper,
  start_mult,
  start_type,
  include_weekends,
  turnaround.time,
  test_start_day,
  test_quarantine,
  vax_eff,
  surveillance,
  rapid_test_sens,
  overdisp_off,
  no_test_vacc,
  rel_trans_HH_symp_child,
  teacher_trans,
  n_contacts_brief = 0,
  rel_trans_brief = 1/50,
  bubble = F,
  version = 2
)
}
\arguments{
\item{N}{number of runs}

\item{n_other_adults}{Number of adults in the school other than primary teachers; defaults to 30}

\item{n_contacts}{Number of sustained contacts outside of the classroom; defaults to 10}

\item{rel_trans_HH}{Relative attack rate of household contact (vs. classrom); defaults to 1}

\item{rel_trans}{Relative attack rate of sustained contact (vs. classroom); defaults to 1/8}

\item{rel_trans_CC}{relative transmission in childcare vs. classroom; defaults to 2}

\item{rel_trans_adult}{relative transmission in staff-staff interactions vs. classroom; defaults to 2}

\item{p_asymp_adult}{Fraction of adults with asymptomatic (unsuspected) disease; defaults to 0.2}

\item{child_prob}{if start_type = "cont", set daily probability of infectious entry for children, defaults to .05}

\item{adult_prob}{if start_type = "cont", set daily probability of infectious entry for adults, defaults to .01}

\item{p_asymp_child}{Fraction of children with asymptomatic (unsuspected) disease; defaults to 0.8}

\item{attack}{Average daily attack rate in adults; defaults to 0.01}

\item{child_trans}{Relative transmissibility of children (vs. adults); defaults to 1}

\item{child_susp}{Relative transmissibility of children (vs. adults); defaults to .5}

\item{child_vax}{Vaccination rate of children; defaults to 0}

\item{p_subclin_adult}{Fraction of adults with subclinical but not techincally asymptomatic disease; defaults to 0}

\item{p_subclin_child}{Fraction of children with subclinical but not techincally asymptomatic disease; defaults to 0}

\item{teacher_susp}{Teacher vaccination rate; defaults to 0.8}

\item{disperse_transmission}{Whether transmission is overdispersed (vs. all have equal attack rate); default to T}

\item{n_staff_contact}{number of contacts a teacher/staff member has with other teachers/staff members; defaults to 1}

\item{n_HH}{number of households a household interacts with when not attending school; defaults to 0}

\item{num_adults}{number of adults interacting with children, defaults to 2}

\item{family_susp}{Household member vaccination rate; defaults to 0.7}

\item{n_start}{number of infections to seed model; defaults to 1}

\item{time_seed_inf}{time(s) at which to introduce new infectious individuals; defaults to NA and randomly selects one time}

\item{days_inf}{length of infectious period (assuming mild case or quarantined on symptoms)}

\item{mult_asymp}{multiplier on asymptomatic infection for adults; default is 1}

\item{mult_asymp_child}{multiplier on asymptomatic infection for children; default is 1}

\item{seed_asymp}{whether to seed with an asymptomatic case}

\item{isolate}{Whether symptomatic individuals isolate when symptoms emerge; defaults to T}

\item{dedens}{Whether dedensification measures reduce attack rate; defaults to F}

\item{run_specials_now}{Whether special subjects are run; defaults to F}

\item{time}{length of time to run model; defaults to 30}

\item{notify}{whether classrooms are notified and quarantined; defaults to F}

\item{test}{whether there is weekly testing; defaults to F}

\item{test_sens}{test sensitivity; defaults to 0.7}

\item{test_frac}{fraction of school tested; defaults to 0.9}

\item{test_days}{test frequency; "day", "week", "2x_week"; defaults to "week"}

\item{test_type}{group tested; defaults to "all", also allows "staff" and "students"}

\item{quarantine.length}{length of quarantine when someone is infectious; defaults to 10}

\item{quarantine.grace}{length of grace period after which a quarantined class returns not to be "re-quarantined"}

\item{type}{schedule; "base", "On/off", "A/B", "Remote"; defaults to "base"}

\item{total_days}{number of days in school; defaults to 5}

\item{includeFamily}{whether to include family, default = FALSE}

\item{synthpop}{synthetic population; defaults to synthpop based on Maryland elementary school}

\item{class}{make_school object; defaults to NA and will call for each simulation}

\item{n_class}{number of classes per grade}

\item{high_school}{whether to use a high school schedule of random period mixing; defaults to F}

\item{nper}{number of school periods; defaults to 8}

\item{start_mult}{value to indicate relative frequency of adult/child infections; defaults to 1 (adults 2x as likely as kids)}

\item{start_type}{type of seed; default is "mix" (also "adult", "child")}

\item{include_weekends}{if TRUE excludes weekends from additional out-of-school mixing, defaults to F}

\item{turnaround.time}{test turnaround time, default = 3 days}

\item{test_start_day}{day tests are implemented for weekly testing; defaults to 1 = Monday}

\item{test_quarantine}{whether quarantined individuals attend school but are tested daily; defaults to FALSE}

\item{vax_eff}{Vaccine efficacy, defaults to 0.9}

\item{surveillance}{whether surveillance is underway; defaults to F}

\item{rapid_test_sens}{sensitivity of rapid tests, defaults to 80\%}

\item{overdisp_off}{all overdispersion off; defaults to F}

\item{no_test_vacc}{Indicates whether vaccinated individuals are excluded from TTS & screening; defaults to F}

\item{rel_trans_HH_symp_child}{Additional relative attack rate of a symptomatic infected child in the household; defaults to 1}

\item{teacher_trans}{Factor by which teacher transmissibility is reduced due to intervention; defaults to 1}

\item{n_contacts_brief}{Number of brief contacts outside of the classroom; defaults to 20}

\item{rel_trans_brief}{Relative attack rate of brief contact (vs. classroom); defaults to 1/50}

\item{bubble}{whether out-of-school interactions occur with a 'bubble'; defaults to F}

\item{version}{v1 quarantines full cohort in A/B; v2 only sub-cohort; defaults to 2}
}
\description{
Run model multiple times and summarize results
}
