% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm6.R
\name{results}
\alias{results}
\title{Summarize multiple runs}
\usage{
results(out)
}
\arguments{
\item{out}{output from mult_runs}
}
\value{
calc summarizes results from multiple runsf
}
\description{
Summarize multiple runs
}
