% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm6.R
\name{run_care}
\alias{run_care}
\title{Set care-based transmission}
\usage{
run_care(a, df, care_contacts, rel_trans_CC = 2, num_adults = 2)
}
\arguments{
\item{a}{id of infected individual}

\item{df}{school data frame from make_school()}

\item{num_adults}{number of adults interacting with children, defaults to 2}

\item{contacts}{graph of random contacts at time t}
}
\value{
infs id of infected individuals
}
\description{
Determine who is infected at a timestep
from contact with an infected individual out of school
}
