% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm6.R
\name{run_class}
\alias{run_class}
\title{Set class transmission}
\usage{
run_class(a, df, high_school = F, hs.classes = NA)
}
\arguments{
\item{a}{id of infected individual}

\item{df}{school data frame from make_school()}
}
\value{
infs id of infected individuals
}
\description{
Determine who is infected at a timestep
in the same classroom as an infected individual
}
