% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm6.R
\name{run_household}
\alias{run_household}
\title{Set household transmission}
\usage{
run_household(a, df)
}
\arguments{
\item{a}{id of infected individual}

\item{df}{school data frame from make_school()}
}
\value{
infs id of infected individuals
}
\description{
Determine who is infected at a timestep
in the same household as an infected individual
}
