% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm6.R
\name{run_model}
\alias{run_model}
\title{Run model}
\usage{
run_model(
  time = 30,
  notify = F,
  test = F,
  test_days = "week",
  test_sens = 0.7,
  test_frac = 0.9,
  n_staff_contact = 0,
  n_HH = 0,
  n_start = 1,
  days_inf = 6,
  mult_asymp = 1,
  seed_asymp = F,
  time_seed_inf = NA,
  high_school = F,
  nper = 8,
  start_mult = 1,
  start_type = "mix",
  test_type = "all",
  adult_prob = 0.013,
  child_prob = 0.056,
  quarantine.length = 14,
  quarantine.grace = 3,
  rel_trans_CC = 2,
  rel_trans_adult = 2,
  num_adults = 2,
  bubble = F,
  no_weekends = F,
  df,
  sched
)
}
\arguments{
\item{time}{length of time to run model; defaults to 30}

\item{notify}{whether classrooms are notified and quarantined; defaults to F}

\item{test}{whether there is weekly testing; defaults to F}

\item{test_sens}{test sensitivity; defaults to 0.7}

\item{test_frac}{fraction of school tested; defaults to 0.9}

\item{n_staff_contact}{number of contacts a teacher/staff member has with other teachers/staff members; defaults to 1}

\item{n_HH}{number of households a household interacts with when not attending school; defaults to 0}

\item{n_start}{number of infections to seed model; defaults to 1}

\item{days_inf}{length of infectious period (assuming mild case or quarantined on symptoms)}

\item{mult_asymp}{multiplier on asymptomatic infection; default is 1}

\item{seed_asymp}{whether to seed with an asymptomatic case}

\item{time_seed_inf}{time(s) at which to introduce new infectious individuals; defaults to NA and randomly selects one time}

\item{high_school}{whether to use a high school schedule of random period mixing; defaults to F}

\item{nper}{number of school periods; defaults to 8}

\item{start_mult}{value to indicate relative frequency of adult/child infections; defaults to 1 (adults 2x as likely as kids)}

\item{start_type}{type of seed; default is "mix" (also "adult", "child", "cont")}

\item{test_type}{group tested; defaults to "all", also allows "staff" and "students"}

\item{quarantine.length}{length of quarantine when someone is infectious; defaults to 14}

\item{quarantine.grace}{length of grace period after which a quarantined class returns not to be "re-quarantined"}

\item{num_adults}{number of adults interacting with children, defaults to 2}

\item{bubble}{whether out-of-school interactions occur with a 'bubble'; defaults to F}

\item{no_weekends}{if TRUE excludes weekends from additional out-of-school mixing, defaults to F}

\item{df}{school data frame from make_school()}

\item{sched}{schedule data frame from make_schedule()}
}
\value{
df updated df with transmission results

time_seed_inf when the first individual was dropped in

class_quarantine a matrix of class quarantine times

mat a check on if the people who you think are present are actually the ones present
}
\description{
Perform a single model run
}
