% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm6.R
\name{run_staff_rand}
\alias{run_staff_rand}
\title{Set random staff transmission}
\usage{
run_staff_rand(a, df, n_contact, rel_trans_adult = 2)
}
\arguments{
\item{a}{id of infected individual}

\item{df}{school data frame from make_school()}

\item{random_contacts}{graph of random contacts at time t}
}
\value{
infs id of infected individuals
}
\description{
Determine who is infected at a timestep
from random contact between in-school adults
}
