% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm6.R
\name{initialize_school}
\alias{initialize_school}
\title{Initialize school}
\usage{
initialize_school(
  n_contacts = 10,
  n_contacts_brief = 0,
  rel_trans_HH = 1,
  rel_trans_HH_symp_child = 1,
  rel_trans = 1/8,
  rel_trans_brief = 1/50,
  p_asymp_adult = 0.35,
  p_asymp_child = 0.7,
  p_subclin_adult = 0,
  p_subclin_child = 0,
  attack = 0.01,
  child_trans = 1,
  child_susp = 0.5,
  family_susp = 1,
  teacher_trans = 1,
  teacher_susp = 1,
  disperse_transmission = T,
  child_vax = 0,
  isolate = T,
  dedens = T,
  run_specials = F,
  start,
  vax_eff = 0.9,
  notify = T,
  no_test_vacc = F
)
}
\arguments{
\item{n_contacts}{Number of sustained contacts outside of the classroom; defaults to 10}

\item{n_contacts_brief}{Number of brief contacts outside of the classroom; defaults to 0}

\item{rel_trans_HH}{Relative attack rate of household contact (vs. classrom); defaults to 1}

\item{rel_trans_HH_symp_child}{Additional relative attack rate of a symptomatic infected child in the household; defaults to 1}

\item{rel_trans}{Relative attack rate of sustained contact (vs. classroom); defaults to 1/8}

\item{rel_trans_brief}{Relative attack rate of brief contact (vs. classroom); defaults to 1/50}

\item{p_asymp_adult}{Fraction of adults with asymptomatic disease; defaults to 0.4}

\item{p_asymp_child}{Fraction of children with asymptomatic disease; defaults to 0.8}

\item{p_subclin_adult}{Fraction of adults with subclinical but not techincally asymptomatic disease; defaults to 0}

\item{p_subclin_child}{Fraction of children with subclinical but not techincally asymptomatic disease; defaults to 0}

\item{attack}{Average daily attack rate in adults; defaults to 0.01}

\item{child_trans}{Relative transmissibility of children (vs. adults); defaults to 1}

\item{child_susp}{Relative transmissibility of children (vs. adults); defaults to .5}

\item{family_susp}{Factor by which adult transmissibility is reduced due to intervention; defaults to 1}

\item{teacher_trans}{Factor by which teacher transmissibility is reduced due to intervention; defaults to 1}

\item{teacher_susp}{Factor by which teacher transmissibility is reduced due to intervention; defaults to 1}

\item{disperse_transmission}{Whether transmission is overdispersed (vs. all have equal attack rate); default to T}

\item{child_vax}{Vaccination rate of children; defaults to 0}

\item{isolate}{Whether symptomatic individuals isolate when symptoms emerge; defaults to T}

\item{dedens}{Whether dedensification measures reduce attack rate; defaults to F}

\item{run_specials}{Whether special subjects are run; defaults to F}

\item{start}{Data frame from make_school()}

\item{vax_eff}{Vaccine efficacy, defaults to 0.9}

\item{notify}{Whether classrooms are notified following a positive test; defaults to T}

\item{no_test_vacc}{Indicates whether vaccinated individuals are excluded from TTS & screening; defaults to F}
}
\value{
out data frame of child and teacher attributes.
}
\description{
This function takes in a data frame exported by make_school().
It adds epidemiological attributes of the full school community.
}
