% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm6.R
\name{mult_runs}
\alias{mult_runs}
\title{Run model multiple times and summarize results}
\usage{
mult_runs(
  N = 500,
  n_other_adults = 30,
  n_contacts = 10,
  n_contacts_brief = 0,
  rel_trans_HH = 1,
  rel_trans = 1/8,
  rel_trans_brief = 1/50,
  rel_trans_CC = 2,
  rel_trans_adult = 2,
  p_asymp_adult = 0.4,
  child_prob = 0.05,
  adult_prob = 0.01,
  p_asymp_child = 0.8,
  attack = 0.01,
  child_trans = 1,
  child_susp = 0.5,
  p_subclin_adult = 0,
  p_subclin_child = 0,
  teacher_trans = 1,
  teacher_susp = 1,
  disperse_transmission = T,
  n_staff_contact = 0,
  n_HH = 0,
  num_adults = 2,
  n_start = 1,
  time_seed_inf = NA,
  days_inf = 6,
  mult_asymp = 1,
  seed_asymp = F,
  isolate = T,
  dedens = 0,
  run_specials_now = F,
  time = 30,
  notify = F,
  test = F,
  test_sens = 0.7,
  test_frac = 0.9,
  test_days = "week",
  test_type = "all",
  quarantine.length = 10,
  quarantine.grace = 3,
  type = "base",
  total_days = 5,
  includeFamily = T,
  synthpop = synthpop,
  class = NA,
  n_class = 4,
  high_school = F,
  nper = 8,
  start_mult = 1,
  start_type = "mix",
  bubble = F,
  include_weekends = T,
  turnaround.time = 1
)
}
\arguments{
\item{N}{number of runs}

\item{n_other_adults}{Number of adults in the school other than primary teachers; defaults to 30}

\item{n_contacts}{Number of sustained contacts outside of the classroom; defaults to 10}

\item{n_contacts_brief}{Number of brief contacts outside of the classroom; defaults to 20}

\item{rel_trans_HH}{Relative attack rate of household contact (vs. classrom); defaults to 1}

\item{rel_trans}{Relative attack rate of sustained contact (vs. classroom); defaults to 1/8}

\item{rel_trans_brief}{Relative attack rate of brief contact (vs. classroom); defaults to 1/50}

\item{p_asymp_adult}{Fraction of adults with asymptomatic (unsuspected) disease; defaults to 0.2}

\item{child_prob}{if start_type = "cont", set daily probability of infectious entry for children, defaults to .05}

\item{adult_prob}{if start_type = "cont", set daily probability of infectious entry for adults, defaults to .01}

\item{p_asymp_child}{Fraction of children with asymptomatic (unsuspected) disease; defaults to 0.8}

\item{attack}{Average daily attack rate in adults; defaults to 0.01}

\item{child_trans}{Relative transmissibility of children (vs. adults); defaults to 1}

\item{child_susp}{Relative transmissibility of children (vs. adults); defaults to .5}

\item{p_subclin_adult}{Fraction of adults with subclinical but not techincally asymptomatic disease; defaults to 0}

\item{p_subclin_child}{Fraction of children with subclinical but not techincally asymptomatic disease; defaults to 0}

\item{teacher_trans}{Factor by which teacher transmissibility is reduced due to intervention; defaults to 1}

\item{teacher_susp}{Factor by which teacher transmissibility is reduced due to intervention; defaults to 1}

\item{disperse_transmission}{Whether transmission is overdispersed (vs. all have equal attack rate); default to T}

\item{n_staff_contact}{number of contacts a teacher/staff member has with other teachers/staff members; defaults to 1}

\item{n_HH}{number of households a household interacts with when not attending school; defaults to 0}

\item{num_adults}{number of adults interacting with children, defaults to 2}

\item{n_start}{number of infections to seed model; defaults to 1}

\item{time_seed_inf}{time(s) at which to introduce new infectious individuals; defaults to NA and randomly selects one time}

\item{days_inf}{length of infectious period (assuming mild case or quarantined on symptoms)}

\item{mult_asymp}{multiplier on asymptomatic infection; default is 1}

\item{seed_asymp}{whether to seed with an asymptomatic case}

\item{isolate}{Whether symptomatic individuals isolate when symptoms emerge; defaults to T}

\item{dedens}{Whether dedensification measures reduce attack rate; defaults to F}

\item{time}{length of time to run model; defaults to 30}

\item{notify}{whether classrooms are notified and quarantined; defaults to F}

\item{test}{whether there is weekly testing; defaults to F}

\item{test_sens}{test sensitivity; defaults to 0.7}

\item{test_frac}{fraction of school tested; defaults to 0.9}

\item{test_days}{vector indicating days on which students are tested; defaults to Sundays}

\item{test_type}{group tested; defaults to "all", also allows "staff" and "students"}

\item{quarantine.length}{length of quarantine when someone is infectious; defaults to 10}

\item{quarantine.grace}{length of grace period after which a quarantined class returns not to be "re-quarantined"}

\item{type}{"base", "On/off", "A/B", "Remote"; defaults to "base"}

\item{total_days}{number of days in school; defaults to 5}

\item{includeFamily}{whether to include family, default = FALSE}

\item{synthpop}{synthetic population; defaults to synthMaryland}

\item{high_school}{whether to use a high school schedule of random period mixing; defaults to F}

\item{nper}{number of school periods; defaults to 8}

\item{start_mult}{value to indicate relative frequency of adult/child infections; defaults to 1 (adults 2x as likely as kids)}

\item{bubble}{whether out-of-school interactions occur with a 'bubble'; defaults to F}

\item{turnaround.time}{test turnaround time, default = 1 day}

\item{run_specials}{Whether special subjects are run; defaults to F}

\item{@start_type}{type of seed; default is "mix" (also "adult", "child")}
}
\description{
Run model multiple times and summarize results
}
